Added SinglyLinkedList.py (#429)
* Added StonePaperScissor.java

A CUI based Stone paper scissors game. User has to enter a choice and the program will randomly make a choice for itself and then a result is printed based on the rules of the standard stone paper scissors.

* Added SinglyLinkedList.py

This is a perfect implementation of a singly linked list in python.