GSSArray program in Java (#433)
* Create towerOfHanoi.c

Implementation of Tower of Hanoi problem in C language. It is a popular application of recursion.

* Create Gssarray.java

GSSArray is a special type of array which doubles its size when the limit is reached. It also sorts the incoming numbers in ascending order.
It is a great program to implement to understand the working of classes in Java.